Change location from LifecycleManager to ValueInjector
