+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/StageLifecycleManager.java
@@ -49,7 +49,6 @@ class StageLifecycleManager {
 
     void executeAfterStageMethods(boolean fakeExecution) throws Throwable {
         executeLifecycleMethods(afterStageRegister, fakeExecution);
-        checkGuaranteedStatesAreInitialized(); //before or after the other after stage methods?
     }
 
     void executeBeforeStageMethods(boolean fakeExecution) throws Throwable {
@@ -69,33 +68,6 @@ class StageLifecycleManager {
             register.fakeExecution();
         } else {
             register.executeMethods();
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private void checkGuaranteedStatesAreInitialized() throws JGivenMissingGuaranteedScenarioStateException {
-        for (Field field: FieldCache.get(instance.getClass())
-                                    .getFieldsWithAnnotation(ProvidedScenarioState.class, ScenarioState.class)) {
-            if (field.isAnnotationPresent(ProvidedScenarioState.class)) {
-                if (field.getAnnotation(ProvidedScenarioState.class).guaranteed()) {
-                    checkInialized(field);
-                }
-            }
-            if (field.isAnnotationPresent(ScenarioState.class)) {
-                if (field.getAnnotation(ScenarioState.class).guaranteed()) {
-                    checkInialized(field);
-                }
-            }
-        }
-    }
-
-    private void checkInialized(Field field) {
-        Object value = null;
-        try {
-            value = field.get(instance);
-        } catch (IllegalAccessException e) { }
-        if (value == null) {
-            throw new JGivenMissingGuaranteedScenarioStateException(field);
         }
     }
